# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "API for managing contacts stored in MongoDB",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/contacts": {
      "get": {
        "summary": "Get all contacts",
        "description": "Returns a list of all stored contacts.",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "description": "Adds a new contact to the database.",
        "parameters": [
          {
            "in": "body",
            "name": "contact",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "Get contact by ID",
        "description": "Returns a contact by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "404": {
            "description": "Contact not found"
          }
        }
      },
      "put": {
        "summary": "Update a contact by ID",
        "description": "Modifies an existing contact.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "contact",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact by ID",
        "description": "Removes a contact from the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully"
          },
          "404": {
            "description": "Contact not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "favoriteColor": {
          "type": "string"
        },
        "birthday": {
          "type": "string",
          "format": "date"
        }
      }
    }
  }
}